Refactor "Sample" structure to sample.hpp Wavbank : Add stereo output Wavbank : Add "loop switch" Rename "Sample::run" to "Sample::loaded" as it is more meaningful Cleanup some non needed includes Wavbank : patching "trig" input do not autoplay sample on sample select
